AddLineClippedSorted

+++ TOMB5/specific/function_table.h
@@ -17,7 +17,7 @@ void SetFogColor(long r, long g, long b);
 #define AddQuadZBuffer	( *(void(__cdecl**)(D3DTLVERTEX*, short, short, short, short, TEXTURESTRUCT*, long)) 0x00876C08 )
 #define AddTriZBuffer	( *(void(__cdecl**)(D3DTLVERTEX*, short, short, short, TEXTURESTRUCT*, long)) 0x00876C10 )
 #define AddTriSorted	( *(void(__cdecl**)(D3DTLVERTEX*, short, short, short, TEXTURESTRUCT*, long)) 0x00876BFC )
-#define AddLineSorted	( *(void(__cdecl**)(D3DTLVERTEX*, D3DTLVERTEX*, long)) 0x00876C00 )
+#define AddLineSorted	( *(void(__cdecl**)(D3DTLVERTEX*, D3DTLVERTEX*, short)) 0x00876C00 )
 #define IsVisible	( *(bool(__cdecl**)(D3DTLVERTEX*, D3DTLVERTEX*, D3DTLVERTEX*)) 0x00876C14 )
 #define _BeginScene	( *(HRESULT(__cdecl**)()) 0x00876C18 )
 #define _EndScene	( *(HRESULT(__cdecl**)()) 0x00876C0C )
